refact: transaction decoding logic
